+++ src/Exceptions/Display/DisplayTreeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace SleepingOwl\Admin\Exceptions\Display;
+
+class DisplayTreeException extends \Exception
+{
+}