+++ src/Form/Element/NamedFormElement.php
@@ -371,7 +371,11 @@ abstract class NamedFormElement extends FormElement
         $count = count($relations);
 
         if ($count === 1) {
-            return $model->getAttribute($this->getModelAttributeKey());
+            $attribute = $model->getAttribute($this->getModelAttributeKey());
+
+            if (! empty($attribute) or is_null($value)) {
+                return $attribute;
+            }
         }
 
         foreach ($relations as $relation) {
@@ -381,10 +385,16 @@ abstract class NamedFormElement extends FormElement
             }
 
             if ($count === 2) {
-                return $model->getAttribute($relation);
+                $attribute = $model->getAttribute($relation);
+
+                if (! empty($attribute) or is_null($value)) {
+                    return $attribute;
+                }
             }
 
-            throw new LogicException("Can not fetch value for field '{$path}'. Probably relation definition is incorrect");
+            if (is_null($this->getDefaultValue())) {
+                throw new LogicException("Can not fetch value for field '{$path}'. Probably relation definition is incorrect");
+            }
         }
 
         return $value;