+++ src/Form/Element/SelectAjax.php
@@ -5,8 +5,8 @@ namespace SleepingOwl\Admin\Form\Element;
 use Illuminate\Routing\Router;
 use Illuminate\Database\Eloquent\Model;
 use SleepingOwl\Admin\Contracts\Initializable;
-use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 use SleepingOwl\Admin\Contracts\WithRoutesInterface;
+use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 
 class SelectAjax extends Select implements Initializable, WithRoutesInterface
 {
@@ -53,6 +53,8 @@ class SelectAjax extends Select implements Initializable, WithRoutesInterface
     public function setSearchUrl($url)
     {
         $this->search_url = $url;
+
+        return $this;
     }
 
     /**