+++ src/Traits/SelectOptionsFromModel.php
@@ -4,8 +4,8 @@ namespace SleepingOwl\Admin\Traits;
 
 use Illuminate\Support\Collection;
 use Illuminate\Database\Eloquent\Model;
-use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 use SleepingOwl\Admin\Exceptions\Form\Element\SelectException;
+use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 
 trait SelectOptionsFromModel
 {