+++ tests/Admin/Form/FormElementTest.php
@@ -36,7 +36,7 @@ class FormElementTest extends TestCase
      */
     public function test_initializable()
     {
-        \KodiCMS\Assets\Facades\Meta::shouldReceive('loadPackage')->once();
+        //\KodiCMS\Assets\Facades\Meta::shouldReceive('loadPackage')->once();
 
         $this->assertNull(
             $this->getElement()->initialize()